refactor(query): split query constants into own module
Moves MAX_BATCH_INSERT_ROWS/MAX_SYNC_PAGE_SIZE out of library/
into a dedicated constants/ folder (one file per constant), and
fixes the resulting broken SelectQueryBuilder import plus a
file/content name swap left over from the move.
