+++ src/database/classes/QueryDb/constants/MAX_BATCH_INSERT_ROWS.ts
@@ -1,0 +1,1 @@
+export const MAX_BATCH_INSERT_ROWS = 500;
