+++ src/database/classes/QueryDb/constants/MAX_SYNC_PAGE_SIZE.ts
@@ -1,0 +1,1 @@
+export const MAX_SYNC_PAGE_SIZE = 500;
