+++ src/database/classes/QueryDb/constants/index.ts
@@ -1,0 +1,2 @@
+export * from './MAX_BATCH_INSERT_ROWS';
+export * from './MAX_SYNC_PAGE_SIZE';
