+++ src/database/classes/QueryDb/library/index.ts
@@ -1,4 +1,3 @@
 export * from './compileCondition';
 export * from './collectConditionColumns';
 export * from './assertIndexedColumns';
-export * from './constants';
